+++ NEWS.md
@@ -42,6 +42,7 @@ MathJax = {
 * Use `getOption("viewer")` instead of `rstudioapi::viewer()` for positron support
 * `glue::glue()` string is accepted by `format_tt()`. Thanks to @LukasWallrich for report #792 on the `modelsummary` repository.
 * Support Github Flavored Markdown (`gfm`) output. Thanks to @kylebutts for contribution #315. 
+* `theme_tt("rotate")` to rotate tables in LaTeX or Typst.
 
 ### HTML